pyproject.toml
[project]
name = "swarmalators-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pupil-labs-uvc==1.0.0b7",
    "pyserial>=3.5",
    "scipy>=1.15.2",
    "simple-pid>=2.0.1",
]
